State Pattern implementation for Bashing
Mainly changed class: Bashing.java
1. changed path for all the classes related to bashing because of the
type error
2. created getter and setter for newly classified states(constant)
3. changed behave() method, state pattern implemented
4. changed newState() method
