+++ src/engine/src/rabbitescape/engine/behaviours/bashing/BashingUselesslyLeftUp.java
@@ -1,16 +1,16 @@
-package rabbitescape.engine.behaviours.Bashing;
+package rabbitescape.engine.behaviours.bashing;
 
 import rabbitescape.engine.ChangeDescription.State;
 import rabbitescape.engine.World;
 import rabbitescape.engine.Rabbit;
 
-import static rabbitescape.engine.ChangeDescription.State.RABBIT_BASHING_USELESSLY_RIGHT_UP;
+import static rabbitescape.engine.ChangeDescription.State.RABBIT_BASHING_USELESSLY_LEFT_UP;
 
-public class BashingUselesslyRightUp implements BashingInterFace {
+public class BashingUselesslyLeftUp implements BashingInterFace {
 
     @Override
     public State getState() {
-        return RABBIT_BASHING_USELESSLY_RIGHT_UP;
+        return RABBIT_BASHING_USELESSLY_LEFT_UP;
     }
 
     @Override
