+++ src/engine/src/rabbitescape/engine/behaviours/bashing/NotBashing.java
@@ -1,21 +1,19 @@
-package rabbitescape.engine.behaviours.Bashing;
+package rabbitescape.engine.behaviours.bashing;
 
 import rabbitescape.engine.ChangeDescription.State;
 import rabbitescape.engine.World;
 import rabbitescape.engine.Rabbit;
 
-import static rabbitescape.engine.ChangeDescription.State.RABBIT_BASHING_USELESSLY_RIGHT;
-
-public class BashingUselesslyRight implements BashingInterFace {
+public class NotBashing implements BashingInterFace {
 
     @Override
     public State getState() {
-        return RABBIT_BASHING_USELESSLY_RIGHT;
+        return null;
     }
 
     @Override
     public boolean behave(World world, Rabbit rabbit) {
         rabbit.slopeBashHop = false;
-        return true;
+        return false;
     }
 }
